+++ example/lib/widget_position_example.dart
@@ -20,6 +20,8 @@ class _WidgetPositionExampleState extends State<WidgetPositionExample> {
     } else {
       widgetPosition = 'Null position';
     }
+
+    setState(() {});
   }
 
   @override